[CDEC-314] Move Lrc code from block to lrc
