+++ block/cardano-sl-block.cabal
@@ -25,6 +25,7 @@ library
                         Pos.Block.Configuration
                         Pos.Block.Error
                         Pos.Block.Logic
+                        Pos.Block.Lrc
                         Pos.Block.RetrievalQueue
                         Pos.Block.Slog
                         Pos.Block.Types
@@ -39,11 +40,6 @@ library
 
                         Pos.GState.BlockExtra
                         Pos.GState.SanityCheck
-
-                        -- LRC
-                        Pos.Lrc
-                        Pos.Lrc.DB
-                        Pos.Lrc.Genesis
 
                         -- Binary serialization
                         Pos.Binary.Block.Types
@@ -63,15 +59,6 @@ library
                         Pos.Block.Slog.Context
                         Pos.Block.Slog.Logic
                         Pos.Block.Slog.Types
-
-                        -- LRC
-                        Pos.Lrc.Consumers
-                        Pos.Lrc.DB.Leaders
-                        Pos.Lrc.DB.Lrc
-                        Pos.Lrc.DB.Richmen
-                        Pos.Lrc.Logic
-                        Pos.Lrc.Worker
-
 
   build-depends:        QuickCheck
                       , aeson
