+++ block/src/Pos/Block/BlockWorkMode.hs
@@ -15,6 +15,7 @@ import           System.Wlog (WithLogger)
 
 import           Pos.Binary.Class (Bi)
 import           Pos.Block.Configuration (HasBlockConfiguration)
+import           Pos.Block.Lrc (LrcModeFull)
 import           Pos.Block.Network.Types (MsgBlock, MsgGetBlocks, MsgGetHeaders, MsgHeaders)
 import           Pos.Block.RetrievalQueue (BlockRetrievalQueue, BlockRetrievalQueueTag)
 import           Pos.Block.Slog (HasSlogContext)
@@ -22,7 +23,6 @@ import           Pos.Block.Types (LastKnownHeader, LastKnownHeaderTag, RecoveryH
                                   RecoveryHeaderTag)
 import           Pos.Communication.Protocol (Message)
 import           Pos.Core.Context (HasPrimaryKey)
-import           Pos.Lrc (LrcModeFull)
 import           Pos.Recovery.Info (MonadRecoveryInfo)
 import           Pos.Security.Params (SecurityParams)
 import           Pos.Shutdown.Class (HasShutdownContext)
