+++ delegation/src/Pos/Delegation.hs
@@ -3,7 +3,6 @@
 module Pos.Delegation
        ( module Pos.Delegation.Worker
        , module Pos.Delegation.Types
-       , module Pos.Delegation.Lrc
        , module Pos.Delegation.Logic
        , module Pos.Delegation.Listeners
        , module Pos.Delegation.DB
@@ -12,14 +11,11 @@ module Pos.Delegation
        , module Pos.Delegation.Cede
        ) where
 
-import           Pos.Delegation.Worker
-import           Pos.Delegation.Types
-import           Pos.Delegation.Lrc
-import           Pos.Delegation.Logic
-import           Pos.Delegation.Listeners
-import           Pos.Delegation.DB
-import           Pos.Delegation.Configuration
-import           Pos.Delegation.Class
 import           Pos.Delegation.Cede
-
-
+import           Pos.Delegation.Class
+import           Pos.Delegation.Configuration
+import           Pos.Delegation.DB
+import           Pos.Delegation.Listeners
+import           Pos.Delegation.Logic
+import           Pos.Delegation.Types
+import           Pos.Delegation.Worker
