+++ delegation/src/Pos/Delegation/Logic/VAR.hs
@@ -26,8 +26,9 @@ import           Serokell.Util (listJson, mapJson)
 import           System.Wlog (WithLogger, logDebug)
 import           UnliftIO (MonadUnliftIO)
 
-import           Pos.Core (ComponentBlock (..), EpochIndex (..), HasGenesisBlockVersionData, StakeholderId,
-                           addressHash, epochIndexL, gbHeader, headerHash, prevBlockL, siEpoch, HasProtocolMagic)
+import           Pos.Core (ComponentBlock (..), EpochIndex (..), HasGenesisBlockVersionData,
+                           HasProtocolMagic, StakeholderId, addressHash, epochIndexL, gbHeader,
+                           headerHash, prevBlockL, siEpoch)
 import           Pos.Core.Block (Block, mainBlockDlgPayload, mainBlockSlot)
 import           Pos.Crypto (ProxySecretKey (..), shortHashF)
 import           Pos.DB (DBError (DBMalformed), MonadDBRead, SomeBatchOp (..))
@@ -43,8 +44,8 @@ import qualified Pos.Delegation.DB as GS
 import           Pos.Delegation.Logic.Common (DelegationError (..), runDelegationStateAction)
 import           Pos.Delegation.Logic.Mempool (clearDlgMemPoolAction, deleteFromDlgMemPool,
                                                processProxySKHeavyInternal)
-import           Pos.Delegation.Lrc (getDlgRichmen)
 import           Pos.Delegation.Types (DlgBlund, DlgPayload (getDlgPayload), DlgUndo (..))
+import           Pos.Lrc.Consumer.Delegation (getDlgRichmen)
 import           Pos.Lrc.Context (HasLrcContext)
 import           Pos.Lrc.Types (RichmenSet)
 import           Pos.Util (getKeys, _neHead)
