+++ lrc/src/Pos/Lrc.hs
@@ -8,12 +8,9 @@ module Pos.Lrc
        , module Pos.Lrc.DB
        , module Pos.Lrc.Error
        , module Pos.Lrc.Fts
-       , module Pos.Lrc.Logic
        , module Pos.Lrc.Mode
        , module Pos.Lrc.RichmenComponent
        , module Pos.Lrc.Types
-       , module Pos.Lrc.Worker
-       , module Test.Pos.Lrc.Arbitrary
        ) where
 
 import           Pos.Lrc.Consumer
@@ -23,9 +20,6 @@ import           Pos.Lrc.Core
 import           Pos.Lrc.DB
 import           Pos.Lrc.Error
 import           Pos.Lrc.Fts
-import           Pos.Lrc.Logic
 import           Pos.Lrc.Mode
 import           Pos.Lrc.RichmenComponent
 import           Pos.Lrc.Types
-import           Pos.Lrc.Worker
-import           Test.Pos.Lrc.Arbitrary
