+++ lrc/src/Pos/Lrc/Consumer/Ssc.hs
@@ -1,23 +1,23 @@
 {-# LANGUAGE TypeFamilies #-}
 
--- | Richmen computation for the update system.
+-- | Richmen computation for SSC.
 
-module Pos.Update.Lrc
+module Pos.Lrc.Consumer.Ssc
        (
        -- * The 'RichmenComponent' instance
-         RCUs
+         RCSsc
 
        -- * The consumer
-       , usLrcConsumer
+       , sscLrcConsumer
 
        -- * Functions for getting richmen
-       , getUSRichmen
-       , tryGetUSRichmen
+       , getSscRichmen
+       , tryGetSscRichmen
        ) where
 
 import           Universum
 
-import           Pos.Core (EpochIndex, HasGenesisBlockVersionData, bvdUpdateVoteThd,
+import           Pos.Core (EpochIndex, HasGenesisBlockVersionData, bvdMpcThd,
                            genesisBlockVersionData)
 import qualified Pos.DB as DB
 import qualified Pos.Lrc.Consumer as Lrc
@@ -30,14 +30,14 @@ import qualified Pos.Lrc.Types as Lrc
 -- RichmenComponent
 ----------------------------------------------------------------------------
 
--- | A tag for the update system 'RichmenComponent'
-data RCUs
+-- | A tag for the SSC 'RichmenComponent'
+data RCSsc
 
-instance HasGenesisBlockVersionData => RichmenComponent RCUs where
-    type RichmenData RCUs = Lrc.FullRichmenData
-    rcToData = identity
-    rcTag Proxy = "us"
-    rcInitialThreshold Proxy = bvdUpdateVoteThd genesisBlockVersionData
+instance HasGenesisBlockVersionData => RichmenComponent RCSsc where
+    type RichmenData RCSsc = Lrc.RichmenStakes
+    rcToData = snd
+    rcTag Proxy = "ssc"
+    rcInitialThreshold Proxy = bvdMpcThd genesisBlockVersionData
     rcConsiderDelegated Proxy = True
 
 ----------------------------------------------------------------------------
@@ -45,30 +45,30 @@ instance HasGenesisBlockVersionData => RichmenComponent RCUs where
 ----------------------------------------------------------------------------
 
 -- | Consumer will be called on every Richmen computation.
-usLrcConsumer :: (HasGenesisBlockVersionData, DB.MonadGState m, DB.MonadDB m) => Lrc.LrcConsumer m
-usLrcConsumer = Lrc.lrcConsumerFromComponentSimple @RCUs bvdUpdateVoteThd
+sscLrcConsumer :: (DB.MonadGState m, DB.MonadDB m, HasGenesisBlockVersionData) => Lrc.LrcConsumer m
+sscLrcConsumer = Lrc.lrcConsumerFromComponentSimple @RCSsc bvdMpcThd
 
 ----------------------------------------------------------------------------
 -- Getting richmen
 ----------------------------------------------------------------------------
 
--- | Wait for LRC results to become available and then get update system
--- ricmen data for the given epoch.
-getUSRichmen
+-- | Wait for LRC results to become available and then get the list of SSC
+-- ricmen for the given epoch.
+getSscRichmen
     :: (MonadIO m, DB.MonadDBRead m, MonadReader ctx m, Lrc.HasLrcContext ctx, HasGenesisBlockVersionData)
     => Text               -- ^ Function name (to include into error message)
     -> EpochIndex         -- ^ Epoch for which you want to know the richmen
-    -> m Lrc.FullRichmenData
-getUSRichmen fname epoch =
+    -> m Lrc.RichmenStakes
+getSscRichmen fname epoch =
     Lrc.lrcActionOnEpochReason
         epoch
-        (fname <> ": couldn't get US richmen")
-        tryGetUSRichmen
+        (fname <> ": couldn't get SSC richmen")
+        tryGetSscRichmen
 
--- | Like 'getUSRichmen', but doesn't wait and doesn't fail.
+-- | Like 'getSscRichmen', but doesn't wait and doesn't fail.
 --
 -- Returns a 'Maybe'.
-tryGetUSRichmen
+tryGetSscRichmen
     :: (DB.MonadDBRead m, HasGenesisBlockVersionData)
-    => EpochIndex -> m (Maybe Lrc.FullRichmenData)
-tryGetUSRichmen = getRichmen @RCUs
+    => EpochIndex -> m (Maybe Lrc.RichmenStakes)
+tryGetSscRichmen = Lrc.getRichmen @RCSsc
