+++ lrc/src/Pos/Lrc/Consumer/Update.hs
@@ -1,23 +1,23 @@
 {-# LANGUAGE TypeFamilies #-}
 
--- | Richmen computation for SSC.
+-- | Richmen computation for the update system.
 
-module Pos.Ssc.Lrc
+module Pos.Lrc.Consumer.Update
        (
        -- * The 'RichmenComponent' instance
-         RCSsc
+         RCUs
 
        -- * The consumer
-       , sscLrcConsumer
+       , usLrcConsumer
 
        -- * Functions for getting richmen
-       , getSscRichmen
-       , tryGetSscRichmen
+       , getUSRichmen
+       , tryGetUSRichmen
        ) where
 
 import           Universum
 
-import           Pos.Core (EpochIndex, HasGenesisBlockVersionData, bvdMpcThd,
+import           Pos.Core (EpochIndex, HasGenesisBlockVersionData, bvdUpdateVoteThd,
                            genesisBlockVersionData)
 import qualified Pos.DB as DB
 import qualified Pos.Lrc.Consumer as Lrc
@@ -30,14 +30,14 @@ import qualified Pos.Lrc.Types as Lrc
 -- RichmenComponent
 ----------------------------------------------------------------------------
 
--- | A tag for the SSC 'RichmenComponent'
-data RCSsc
+-- | A tag for the update system 'RichmenComponent'
+data RCUs
 
-instance HasGenesisBlockVersionData => RichmenComponent RCSsc where
-    type RichmenData RCSsc = Lrc.RichmenStakes
-    rcToData = snd
-    rcTag Proxy = "ssc"
-    rcInitialThreshold Proxy = bvdMpcThd genesisBlockVersionData
+instance HasGenesisBlockVersionData => RichmenComponent RCUs where
+    type RichmenData RCUs = Lrc.FullRichmenData
+    rcToData = identity
+    rcTag Proxy = "us"
+    rcInitialThreshold Proxy = bvdUpdateVoteThd genesisBlockVersionData
     rcConsiderDelegated Proxy = True
 
 ----------------------------------------------------------------------------
@@ -45,30 +45,30 @@ instance HasGenesisBlockVersionData => RichmenComponent RCSsc where
 ----------------------------------------------------------------------------
 
 -- | Consumer will be called on every Richmen computation.
-sscLrcConsumer :: (DB.MonadGState m, DB.MonadDB m, HasGenesisBlockVersionData) => Lrc.LrcConsumer m
-sscLrcConsumer = Lrc.lrcConsumerFromComponentSimple @RCSsc bvdMpcThd
+usLrcConsumer :: (HasGenesisBlockVersionData, DB.MonadGState m, DB.MonadDB m) => Lrc.LrcConsumer m
+usLrcConsumer = Lrc.lrcConsumerFromComponentSimple @RCUs bvdUpdateVoteThd
 
 ----------------------------------------------------------------------------
 -- Getting richmen
 ----------------------------------------------------------------------------
 
--- | Wait for LRC results to become available and then get the list of SSC
--- ricmen for the given epoch.
-getSscRichmen
+-- | Wait for LRC results to become available and then get update system
+-- ricmen data for the given epoch.
+getUSRichmen
     :: (MonadIO m, DB.MonadDBRead m, MonadReader ctx m, Lrc.HasLrcContext ctx, HasGenesisBlockVersionData)
     => Text               -- ^ Function name (to include into error message)
     -> EpochIndex         -- ^ Epoch for which you want to know the richmen
-    -> m Lrc.RichmenStakes
-getSscRichmen fname epoch =
+    -> m Lrc.FullRichmenData
+getUSRichmen fname epoch =
     Lrc.lrcActionOnEpochReason
         epoch
-        (fname <> ": couldn't get SSC richmen")
-        tryGetSscRichmen
+        (fname <> ": couldn't get US richmen")
+        tryGetUSRichmen
 
--- | Like 'getSscRichmen', but doesn't wait and doesn't fail.
+-- | Like 'getUSRichmen', but doesn't wait and doesn't fail.
 --
 -- Returns a 'Maybe'.
-tryGetSscRichmen
+tryGetUSRichmen
     :: (DB.MonadDBRead m, HasGenesisBlockVersionData)
-    => EpochIndex -> m (Maybe Lrc.RichmenStakes)
-tryGetSscRichmen = Lrc.getRichmen @RCSsc
+    => EpochIndex -> m (Maybe Lrc.FullRichmenData)
+tryGetUSRichmen = getRichmen @RCUs
