+++ lrc/test/LICENSE
