+++ lrc/test/Test/Pos/Lrc/Arbitrary.hs
