+++ lrc/test/cardano-sl-lrc-test.cabal
