+++ ssc/Pos/Ssc/Logic/Global.hs
@@ -12,12 +12,13 @@ import           System.Wlog (WithLogger)
 import           Universum
 
 import           Pos.Binary.Ssc ()
-import           Pos.Core (EpochIndex (..), SharedSeed, VssCertificatesMap (..), vcVssKey, HasGenesisBlockVersionData)
+import           Pos.Core (EpochIndex (..), HasGenesisBlockVersionData, SharedSeed,
+                           VssCertificatesMap (..), vcVssKey)
 import           Pos.DB (MonadDBRead)
+import           Pos.Lrc.Consumer.Ssc (getSscRichmen)
 import           Pos.Lrc.Context (HasLrcContext)
 import           Pos.Lrc.Types (RichmenStakes)
 import           Pos.Ssc.Error (SscSeedError)
-import           Pos.Ssc.Lrc (getSscRichmen)
 import           Pos.Ssc.Mem (MonadSscMem, SscGlobalQuery, sscRunGlobalQuery)
 import           Pos.Ssc.Seed (calculateSeed)
 import           Pos.Ssc.Types (sgsCommitments, sgsOpenings, sgsShares, sgsVssCertificates)
