+++ ssc/Pos/Ssc/Worker.hs
@@ -23,12 +23,12 @@ import           Pos.Arbitrary.Ssc ()
 import           Pos.Binary.Class (AsBinary, Bi, asBinary, fromBinary)
 import           Pos.Binary.Infra ()
 import           Pos.Binary.Ssc ()
-import           Pos.Core (EpochIndex, SlotId (..), StakeholderId, Timestamp (..),
-                           VssCertificate (..), VssCertificatesMap (..), blkSecurityParam,
-                           bvdMpcThd, getOurSecretKey, getOurStakeholderId, getSlotIndex, lookupVss,
-                           memberVss, mkLocalSlotIndex, mkVssCertificate, slotSecurityParam,
-                           vssMaxTTL, HasProtocolConstants, HasProtocolMagic, HasGenesisData,
-                           HasGenesisBlockVersionData)
+import           Pos.Core (EpochIndex, HasGenesisBlockVersionData, HasGenesisData,
+                           HasProtocolConstants, HasProtocolMagic, SlotId (..), StakeholderId,
+                           Timestamp (..), VssCertificate (..), VssCertificatesMap (..),
+                           blkSecurityParam, bvdMpcThd, getOurSecretKey, getOurStakeholderId,
+                           getSlotIndex, lookupVss, memberVss, mkLocalSlotIndex, mkVssCertificate,
+                           slotSecurityParam, vssMaxTTL)
 import           Pos.Core.Ssc (Commitment (..), InnerSharesMap, Opening, SignedCommitment,
                                getCommitmentsMap)
 import           Pos.Crypto (SecretKey, VssKeyPair, VssPublicKey, randomNumber, runSecureRandom)
@@ -36,6 +36,7 @@ import           Pos.Crypto.Configuration (protocolMagic)
 import           Pos.Crypto.SecretSharing (toVssPublicKey)
 import           Pos.DB (gsAdoptedBVData)
 import           Pos.Diffusion.Types (Diffusion (..))
+import           Pos.Lrc.Consumer.Ssc (getSscRichmen)
 import           Pos.Lrc.Types (RichmenStakes)
 import           Pos.Recovery.Info (recoveryCommGuard)
 import           Pos.Reporting.MemState (HasMisbehaviorMetrics (..), MisbehaviorMetrics (..))
@@ -48,7 +49,6 @@ import           Pos.Ssc.Configuration (HasSscConfiguration, mpcSendInterval)
 import           Pos.Ssc.Functions (hasCommitment, hasOpening, hasShares, vssThreshold)
 import           Pos.Ssc.Logic (sscGarbageCollectLocalData, sscProcessCertificate,
                                 sscProcessCommitment, sscProcessOpening, sscProcessShares)
-import           Pos.Ssc.Lrc (getSscRichmen)
 import           Pos.Ssc.Message (SscTag (..))
 import           Pos.Ssc.Mode (SscMode)
 import qualified Pos.Ssc.SecretStorage as SS
