+++ update/Pos/Update/Poll/DBPoll.hs
@@ -18,12 +18,12 @@ import           UnliftIO (MonadUnliftIO)
 
 import           Pos.Core (Coin, HasGenesisBlockVersionData)
 import           Pos.DB.Class (MonadDBRead)
+import           Pos.Lrc.Consumer.Update (tryGetUSRichmen)
 import           Pos.Lrc.Context (HasLrcContext, lrcActionOnEpochReason)
 import           Pos.Lrc.DB.Issuers (getIssuersStakes)
 import           Pos.Lrc.Types (FullRichmenData)
 import           Pos.Update.Configuration (HasUpdateConfiguration)
 import qualified Pos.Update.DB as GS
-import           Pos.Update.Lrc (tryGetUSRichmen)
 import           Pos.Update.Poll.Class (MonadPollRead (..))
 
 ----------------------------------------------------------------------------
